Use f32s for glyph metric fractional precision

+++ engine/font.rs
@@ -56,19 +56,19 @@ impl Font {
 			let position_y = u32::from_ne_bytes(bytes) as f32;
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 12..glyph_offset + 16]);
-			let width = u32::from_ne_bytes(bytes) as f32;
+			let width = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 16..glyph_offset + 20]);
-			let height = u32::from_ne_bytes(bytes) as f32;
+			let height = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 20..glyph_offset + 24]);
-			let bearing_x = i32::from_ne_bytes(bytes) as f32;
+			let bearing_x = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 24..glyph_offset + 28]);
-			let bearing_y = i32::from_ne_bytes(bytes) as f32;
+			let bearing_y = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 28..glyph_offset + 32]);
-			let advance = i32::from_ne_bytes(bytes) as f32;
+			let advance = f32::from_ne_bytes(bytes);
 
 			glyphs.push(Glyph {
 				char_code,
